edit card Title and add file css from Jiyeon
