+++ src/main/java/kr/or/bit/dao/CardDao.java
@@ -4,6 +4,7 @@ import java.util.ArrayList;
 import java.util.HashMap;
 
 import kr.or.bit.dto.card;
+import kr.or.bit.dto.list;
 
 public interface CardDao {
 	
@@ -16,4 +17,6 @@ public interface CardDao {
 	public int getCardSeq(int projectNum);
 	
 	public void deleteKanbanCard(int cardSeq);
+	
+	public void editKanbanCardTitle(card ca);
 }
