+++ src/main/webapp/resources/css/kanban-board.css
@@ -260,25 +260,6 @@ html, body {
     background-color: #f8df8c;
 }
 
-#cardTitleForm {
-    height: 100%;
-    float: left;
-    line-height:50px;
-    display: none;
-}
-
-#cardTitleInput {
-    font-family: 'Lato', sans-serif;
-    font-size: 17px;
-    background: transparent;
-    color: #000;
-    border: none;
-    background: #f5df4d;
-    border-radius: 5px;
-    height: 35px;
-    padding-left: 10px;
-    font-weight: bolder;
-}
 /*Add a List*/
 #addListWrap {
     display: inline-block;
