+++ src/main/webapp/resources/css/kanbanDetailModal.css
@@ -1,6 +1,6 @@
 /*
 파일명: checkList.js
-설명: 칸반 카드 상세페이지 내 구현될 체크리스트 css
+설명: 칸반 카드 모달 상세페이지 내 구현될 카드 정보, 체크리스트 css
 작성일: 2021-01-04 ~ 
 작성자: 문지연
 */
@@ -25,7 +25,7 @@
     margin: auto;
     padding: 20px;
     border: 1px solid #888;
-    width: 60%;
+    width: 40%;
     height: 65%;
     border-radius: 15px;
 }
@@ -52,31 +52,34 @@
 
 /*card Title in Modal*/
 .cardTitleMo {
-    font-weight: bolder;
+	font-size:30px;
     padding-left:10px;
-    margin-bottom:20px;
+    margin-bottom:10px;
     width:30%;
+    font-weight: 200;
+	color: #454545;
 }
 
 #cardTitleForm {
-    height: 100%;
-    float: left;
+    width: 30%;
+    padding-left:10px;
+    margin-bottom: 10px;
     line-height:50px;
     display: none;
 }
 
 #cardTitleInput {
-    font-family: 'Lato', sans-serif;
-    font-size: 17px;
+	font-size:28px;
+    padding-left:10px;
+    margin-bottom:10px;
+    font-weight: 200;
     background: transparent;
     color: #000;
     border: none;
     background: #f5df4d;
     border-radius: 5px;
-    height: 35px;
-    padding-left: 10px;
-    font-weight: bolder;
 }
+
 /*Check List in Modal*/
 #checkListWrap {
 	/*transition*/
@@ -95,7 +98,6 @@
 	text-align: left;
 	padding: 20px 0;
 	font-size: 30px;
-	margin: 0;
 	font-weight: 200;
 	color: #454545;
 }
@@ -107,7 +109,7 @@
 }
 
 #checkListFrom {
-	width : 500px;
+	width : 80%;
 	margin: 20px auto 20px 5px;
 	padding: 50px;
 	background: white;
@@ -120,7 +122,6 @@
 	-webkit-border-radius: 5px;
 	-moz-border-radius: 5px;
 	border-radius: 5px;
-	width: 500px;
 }
 
 #checkListFrom:before {
