+++ src/main/webapp/resources/js/kanbanDetailModal.js
@@ -5,8 +5,71 @@
 작성자: 문지연
 */
 
-/*CheckList in Modal*/
 $(document).ready(function(){
+	
+	const details = document.getElementById("detailsModal");
+
+	//get card Title in Modal
+	$(document).on("click", ".cardContent", function(e) {
+		e.preventDefault();
+		$('.modal_card_seq').attr("value",$(this).data().cardSeq);
+		$('.cardTitleMo').text($(this).context.innerText);
+		details.style.display = "block";
+	});
+
+	//close Modal
+	$(document).on("click", ".closeModal", function(e) {
+		e.preventDefault();
+		details.style.display = "none";
+	});
+
+	window.onclick = function(e) {
+		if (e.target == details) {
+			details.style.display = "none";
+		}
+	}
+
+	//open edit card Title form in Modal
+	$(document).on("click", ".cardTitleMo", function(e) {
+		e.preventDefault();
+		let cardTitleForm = $(this).parent().children("#cardTitleForm");
+		$(this).hide();
+		cardTitleForm.children("#cardTitleInput").attr("placeholder", $(this).html());
+		cardTitleForm.show();
+		cardTitleForm.focus();
+	});
+	
+	//edit card Title in Modal
+	$(document).on("submit",".cardTitleForm",function(e){
+		e.preventDefault();
+		let editedCardTitle = $(this).parents().children().children("#cardTitleInput").val();
+		let modal_card_seq = $(this).parents().children().children('.modal_card_seq').attr('value');
+
+		let cardOb = new Object();
+		cardOb.card_seq = modal_card_seq;
+		cardOb.card_name = editedCardTitle;
+
+		let card = JSON.stringify(cardOb);
+		if (editedCardTitle.length > 0) {
+			$.ajax({
+				type: "post",
+				url: "editKanbanCardTitle.pie",
+				contentType: "application/json; charset=UTF-8",
+				dataType: "json",
+				async: false,
+				data: card,
+				success: function(data) {
+					console.log(data);
+				}
+			});
+			$('.cardTitleMo').text(editedCardTitle);
+			$("[data-card-seq="+modal_card_seq+"]").text(editedCardTitle);
+			$('.cardTitleForm').hide();
+			$('.cardTitleMo').show();
+		}
+	});
+	
+	/*CheckList in Modal*/
 	$('#add-todo').click(function() {
 		const lastSibling = $('#checkListFrom > .todo-wrap:last-of-type > input').attr('id');
 		const newId = Number(lastSibling) + 1;
